Decompose launcher_core lib.rs into cohesive modules
lib.rs (1,144 lines) becomes options (public config types), setup
(first-run config + secret files), stack (process spawn/launch/stop),
process (instance lock, stale child cleanup, termination), proxy
(local HTTP proxy + frontend resolution), tools (ffmpeg/ffprobe
discovery), and util. lib.rs keeps the public API via re-exports;
standalone_launcher and the desktop app compile unchanged.
